I re-wrote alot of things which I didn't like, and I tried to use the 'left join' concept we talked about (AG/ML) some time ago. In "short": a report has one "line" table, and can have many "column" tables, each of which include a list of fields. Filters are then applied.
The tricky part, is that some fields cause enumerations (lcm_fields.enum)
which makes left joining really difficult (or incredebly SQL intensive, I
tested many days on a few days.. I lost hope). So the SQL queries for those
fields are included in SQL headers, which are later interpreted. A good
example of this, is to list "number of FU per case", where you have a list
of cases, and one column for each type of FU. I will document more later.